Prototyping/Signals/pyproject.toml
[project]
name = "external-signals-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "beautifulsoup4>=4.14.3",
    "builtwith>=1.3.4",
    "db-dtypes>=1.5.0",
    "google-cloud-bigquery>=3.40.0",
    "google-generativeai>=0.8.6",
    "lxml>=6.0.2",
    "nltk>=3.9.2",
    "pandas>=2.3.3",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.1",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jobspy>=1.1.82",
]
